LayoutTests/inspector/elements/iframe-load-event.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script src="resources/iframe-load-event-iframe.js"></script>
</head>
<body>
<p>
Tests that iframe content is available after iframe's load event fired. See <a href="http://webkit.org/b/76552">bug 76552</a>.
</p>
<iframe id="myframe" src="resources/iframe-load-event-iframe-1.html" onload="runTest()"></iframe>
</body>
</html>
