LayoutTests/inspector/elements/iframe-load-event-expected.txt

Tests that iframe content is available after iframe's load event fired. See bug 76552.




After frame navigate
- <html>
    - <head>
          <script src="../../http/tests/inspector/inspector-test.js"></script>
          <script src="../../http/tests/inspector/elements-test.js"></script>
          <script src="resources/iframe-load-event-iframe.js"></script>
      </head>
    - <body>
        - <p>
              "\nTests that iframe content is available after iframe's load event fired. See "
              <a href="http://webkit.org/b/76552">bug 76552</a>
              ".\n"
          </p>
        - <iframe id="myframe" src="resources/iframe-load-event-iframe-2.html" onload="runTest()">
            - #document
            - <html>
                  <head></head>
                - <body>
                      <div id="iframe-2-element"></div>
                  </body>
              </html>
          </iframe>
      </body>
  </html>
